+++ package.json
@@ -11,6 +11,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1",
+    "swr": "^2.2.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
